+++ webapp/src/components/MarkdownEditor/index.jsx
@@ -1,14 +1,15 @@
+import '@toast-ui/editor/dist/toastui-editor.css';
 import React, { useCallback, useRef } from 'react';
 import PropTypes from 'prop-types';
-import '@toast-ui/editor/dist/toastui-editor.css';
 import { Editor, Viewer } from '@toast-ui/react-editor';
 
 MarkdownEditor.propTypes = {
   mdValue: PropTypes.string.isRequired,
   setContent: PropTypes.func.isRequired,
+  height: PropTypes.string,
 };
 
-function MarkdownEditor({ mdValue, setContent }) {
+export default function MarkdownEditor({ mdValue, setContent, height = '300px' }) {
   const editorRef = useRef(null);
   const onChangeEditorTextHandler = useCallback(() => {
     setContent(editorRef.current?.getInstance().getMarkdown());
@@ -19,9 +20,8 @@ function MarkdownEditor({ mdValue, setContent }) {
       initialEditType="markdown"
       useCommandShortcut
       ref={editorRef}
+      height={height}
       onChange={onChangeEditorTextHandler}
     />
   );
 }
-
-export default React.memo(MarkdownEditor);
