+++ webapp/src/pages/NewTeamPost/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
-import MarkdownEditor from 'components/MdEditor';
+import MarkdownEditor from 'components/MarkdownEditor';
 import useFileUploader from 'hooks/useFileInput';
 import useInput from 'hooks/useInput';
 import { hopeSessionOption, skillStack } from 'constant';
